requirements.txt
# Housing Compliance Web Application Requirements

# Core Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# HTTP Requests (for API calls)
requests==2.31.0

# Image Processing (for validation engine)
Pillow==10.1.0

# JSON Processing
# (Built-in json module is sufficient)

# Development Dependencies (optional)
# flask-cors==4.0.0  # For CORS if needed
# python-dotenv==1.0.0  # For environment variables